+++ app/src/main/java/eu/kanade/presentation/manga/components/BaseMangaListItem.kt
@@ -18,7 +18,6 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import tachiyomi.domain.manga.model.Manga
 import tachiyomi.presentation.core.components.material.padding
-import tachiyomi.presentation.core.util.selectedBackground
 
 @Composable
 fun BaseMangaListItem(
@@ -28,7 +27,6 @@ fun BaseMangaListItem(
     onClickCover: () -> Unit = onClickItem,
     // KMK -->
     onLongClick: () -> Unit = onClickItem,
-    selected: Boolean,
     // KMK <--
     cover: @Composable RowScope.() -> Unit = { defaultCover(manga, onClickCover) },
     actions: @Composable RowScope.() -> Unit = {},
@@ -40,7 +38,6 @@ fun BaseMangaListItem(
     Row(
         modifier = modifier
             // KMK -->
-            .selectedBackground(selected)
             .combinedClickable(
                 // KMK <--
                 onClick = onClickItem,
